Minor refactoring in a view of C#-porting
Autoported commit.
Original commit hash: [4d0f2fe]

+++ itext/itext.layout/itext/layout/renderer/ImageRenderer.cs
@@ -152,8 +152,8 @@ namespace iText.Layout.Renderer {
                     return new LayoutResult(LayoutResult.NOTHING, occupiedArea, null, this, this);
                 }
             }
-            occupiedArea.GetBBox().MoveDown(height);
-            occupiedArea.GetBBox().SetHeight(height);
+            occupiedArea.GetBBox().MoveDown((float)height);
+            occupiedArea.GetBBox().SetHeight((float)height);
             occupiedArea.GetBBox().SetWidth((float)width);
             float leftMargin = (float)this.GetPropertyAsFloat(Property.MARGIN_LEFT);
             float topMargin = (float)this.GetPropertyAsFloat(Property.MARGIN_TOP);
@@ -257,9 +257,9 @@ namespace iText.Layout.Renderer {
             if (angle != 0) {
                 AffineTransform t = AffineTransform.GetRotateInstance(angle);
                 Point p00 = t.Transform(new Point(0, 0), new Point());
-                Point p01 = t.Transform(new Point(0, height), new Point());
+                Point p01 = t.Transform(new Point(0, (float)height), new Point());
                 Point p10 = t.Transform(new Point((float)width, 0), new Point());
-                Point p11 = t.Transform(new Point((float)width, height), new Point());
+                Point p11 = t.Transform(new Point((float)width, (float)height), new Point());
                 double[] xValues = new double[] { p01.GetX(), p10.GetX(), p11.GetX() };
                 double[] yValues = new double[] { p01.GetY(), p10.GetY(), p11.GetY() };
                 double minX = p00.GetX();
@@ -285,7 +285,7 @@ namespace iText.Layout.Renderer {
             float scaleCoeff = 1;
             // hasProperty(Property) checks only properties field, cannot use it
             if (true.Equals(GetPropertyAsBoolean(Property.AUTO_SCALE))) {
-                scaleCoeff = Math.Min(maxWidth / (float)width, maxHeight / height);
+                scaleCoeff = Math.Min(maxWidth / (float)width, maxHeight / (float)height);
                 height *= scaleCoeff;
                 width *= scaleCoeff;
             }
@@ -299,7 +299,7 @@ namespace iText.Layout.Renderer {
                 else {
                     if (null != GetPropertyAsBoolean(Property.AUTO_SCALE_HEIGHT) && (bool)GetPropertyAsBoolean(Property.AUTO_SCALE_HEIGHT
                         )) {
-                        scaleCoeff = maxHeight / height;
+                        scaleCoeff = maxHeight / (float)height;
                         height = maxHeight;
                         width *= scaleCoeff;
                     }
